ApplicationTest : Adapt for Windows
- There is no known way of renaming the python / cmd.exe process, so we
test against that until a solution is found.
- Killing the `timeout` process on Windows requires running the command
`TASKKILL`.
- Use Python to check environment variables. The `set` command on
Windows is the equivalent to Linux `env` command. But it is a built-in
shell command, which we can't run from Gaffer's `env` command because
we aren't using the shell.

+++ python/GafferTest/ApplicationTest.py
@@ -55,25 +55,56 @@ class ApplicationTest( GafferTest.TestCase ) :
 
 	def testWrapperDoesntDuplicatePaths( self ) :
 
-		output = subprocess.check_output( [ "gaffer", "env", "env" ], universal_newlines = True )
-		externalEnv = {}
-		for line in output.split( '\n' ) :
-			partition = line.partition( "=" )
-			externalEnv[partition[0]] = partition[2]
+		executable = "gaffer" if os.name != "nt" else "gaffer.cmd"
 
-		self.assertEqual( externalEnv["GAFFER_STARTUP_PATHS"], os.environ["GAFFER_STARTUP_PATHS"] )
-		self.assertEqual( externalEnv["GAFFER_APP_PATHS"], os.environ["GAFFER_APP_PATHS"] )
+		for v in ["GAFFER_STARTUP_PATHS", "GAFFER_APP_PATHS"] :
+			value = subprocess.check_output( [ executable, "env", "python", "-c", "import os; print(os.environ['{}'])".format( v ) ], universal_newlines = True )
+			self.assertEqual( value.strip(), os.environ[v] )
 
 	def testProcessName( self ) :
 
-		process = subprocess.Popen( [ "gaffer", "env", "sleep", "100" ] )
-		time.sleep( 1 )
-		command = subprocess.check_output( [ "ps", "-p", str( process.pid ), "-o", "command=" ], universal_newlines = True ).strip()
-		name = subprocess.check_output( [ "ps", "-p", str( process.pid ), "-o", "comm=" ], universal_newlines = True ).strip()
-		process.kill()
+		if os.name == "nt" :
+			process = subprocess.Popen( [ "gaffer.cmd", "env", "python", "-c", "import time; time.sleep(5)" ] )
 
-		self.assertEqual( command, "gaffer env sleep 100" )
-		self.assertEqual( name, "gaffer" )
+			time.sleep( 1 )
+
+			command = subprocess.check_output(
+				[
+					"powershell",
+					"-command",
+					"Get-WmiObject -Query \"SELECT CommandLine FROM Win32_Process WHERE ProcessID={}\" | Format-List -Property CommandLine".format( process.pid )
+				],
+				universal_newlines = True
+			)
+			command = " ".join( [ i.strip() for i in command.strip().split( "\n" ) ] )
+			command = command.replace( "CommandLine : ", "" )
+
+			name = subprocess.check_output(
+				[
+					"powershell",
+					"-command",
+					"Get-WmiObject -Query \"SELECT Name FROM Win32_Process WHERE ProcessID={}\" | Format-List -Property Name".format( process.pid )
+				],
+				universal_newlines = True
+			)
+			name = name.strip().replace( "Name : ", "" )
+
+			subprocess.check_call( [ "TASKKILL", "/F", "/PID", str( process.pid ), "/T" ] )
+
+			self.assertEqual( command, "C:\\Windows\\system32\\cmd.exe /c gaffer.cmd env python -c \"import time; time.sleep(5)\"" )
+			self.assertEqual( name, "cmd.exe" )
+
+		else :
+			process = subprocess.Popen( [ "gaffer", "env", "sleep", "100" ] )
+			time.sleep( 1 )
+
+			command = subprocess.check_output( [ "ps", "-p", str( process.pid ), "-o", "command=" ], universal_newlines = True ).strip()
+			name = subprocess.check_output( [ "ps", "-p", str( process.pid ), "-o", "comm=" ], universal_newlines = True ).strip()
+
+			process.kill()
+
+			self.assertEqual( command, "gaffer env sleep 100" )
+			self.assertEqual( name, "gaffer" )
 
 if __name__ == "__main__":
 	unittest.main()
